requirements.txt
marker-pdf
ollama
torch
python-dotenv